refactor(owners): add `is_valid_owner`
